Index.js to RankingEventComponent
